requirements.txt
streamlit
psycopg2-binary
sqlalchemy
requests
pandas
openpyxl
scikit-learn
langchain
langchain-community
langchain-openai
langchain-core
torch
torchvision
torchaudio
transformers
datasets
accelerate
chromadb
pysqlite3-binary
fastapi
uvicorn
pydantic
python-dotenv
httpx

langchain-groq
langchain-google-genai
psycopg2-binary

deepeval
groq